% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_dates}
\alias{create_dates}
\title{create_dates
Creates a tibble with a column of start dates and column of end dates by year}
\usage{
create_dates(start_date, end_date)
}
\arguments{
\item{start_date}{Date}

\item{end_date}{Date}
}
\value{
tibble
}
\description{
create_dates
Creates a tibble with a column of start dates and column of end dates by year
}
\keyword{internal}
