% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_ncdc}
\alias{download_ncdc}
\title{download_ncdc}
\usage{
download_ncdc(start_date, end_date, station_id, data_type_id, token,
  progress = TRUE)
}
\arguments{
\item{start_date}{Date}

\item{end_date}{Date}

\item{station_id}{character}

\item{data_type_id}{character}

\item{token}{character}

\item{progress}{logical}
}
\value{
tibble
}
\description{
Functionally downloads specified data using rnoaa's ncdc function over the specified date range. Use this when you have to download multiple years worth of data
}
