% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data}
\alias{get_data}
\title{get_data}
\usage{
get_data(start_date, end_date, data_type_id, station_id, token,
  add_units = TRUE)
}
\arguments{
\item{start_date}{Date}

\item{end_date}{Date}

\item{data_type_id}{character}

\item{station_id}{character}

\item{token}{character}

\item{add_units}{logical}
}
\value{
dataframe
}
\description{
get_data
}
\keyword{internal}
